Move babel-plugin-strip-runtime stylesheet extraction tests to own file
